cangoprocess/requirements.txt
azure-functions
azure-ai-formrecognizer
azure-cosmos
pandas